Add files via upload
I added grass textures to the blocks. Due to time constraints, I was unable to make a full map with all three textures as intended. However, I did include all three textures in this commit anyway. I believe for it to run, the three texture jpgs need to be in a folder called "textures."